refactor main script into service package
